refactor: 각 메뉴 마다  랜더링 하는 함수 분리
