+++ src/js/utils/dom.js
@@ -1,1 +1,6 @@
 export const $ = (select) => document.querySelector(select);
+
+export const replaceElement = (element, content) => {
+  element.replaceChildren();
+  element.insertAdjacentHTML('beforeend', content);
+};
